feat(template): update git repo in instrumentation hook
